+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 1.1.0 - Dec 11, 2020
+- Added `getcount` command.
+	- Displays total emoji count usage at all-time, monthly, and weekly usage; defaults to server's usage if user is not specified.
+
 ### 1.0.1 - Dec 3, 2020
 - Added next/prev page reaction removal on timeout for the `displaystats` command.
 - Fixed bug for next/prev page reaction removal.